requirements.txt
# Core dependencies
fastapi>=0.100.0
pydantic>=2.1.1
uvicorn>=0.23.1

# Vector store
faiss-cpu>=1.7.4
numpy>=1.25.1

# Real-time pipeline
pathway>=0.4.1

# LLM & embeddings
openai==1.88.0

# Optional - live news feed
requests>=2.31.0

# Testing
pytest>=7.4.0